+++ src/Symfony/Component/Security/Core/Event/VoteEvent.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Security\Core\Event;
 
+use Symfony\Component\Security\Core\Authorization\Voter\Vote;
+use Symfony\Component\Security\Core\Authorization\Voter\VoteInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 use Symfony\Contracts\EventDispatcher\Event;
 
@@ -23,12 +25,15 @@ use Symfony\Contracts\EventDispatcher\Event;
  */
 final class VoteEvent extends Event
 {
+    private VoteInterface $vote;
+
     public function __construct(
         private VoterInterface $voter,
         private mixed $subject,
         private array $attributes,
-        private int $vote,
+        VoteInterface|int $vote,
     ) {
+        $this->vote = $vote instanceof VoteInterface ? $vote : new Vote($vote);
     }
 
     public function getVoter(): VoterInterface
@@ -47,6 +52,11 @@ final class VoteEvent extends Event
     }
 
     public function getVote(): int
+    {
+        return $this->vote->getAccess();
+    }
+
+    public function getVoteObject(): VoteInterface
     {
         return $this->vote;
     }