+++ src/Symfony/Component/Security/Core/Tests/Fixtures/DummyVoter.php
@@ -12,11 +12,16 @@
 namespace Symfony\Component\Security\Core\Tests\Fixtures;
 
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Authorization\Voter\Vote;
 use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 
 final class DummyVoter implements VoterInterface
 {
     public function vote(TokenInterface $token, $subject, array $attributes): int
+    {
+    }
+
+    public function getVote(TokenInterface $token, mixed $subject, array $attributes): VoterInterface
     {
     }
 }